+++ index.html
@@ -231,7 +231,7 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <scrypt src="/js/function.js"></scrypt>
-  <scrypt src="/js/main.js"></scrypt>
+  <scrypt src="js/function.js"></scrypt>
+  <scrypt src="js/main.js" type="module"></scrypt>
 </body>
 </html>
